allow custom email address instead of username
In #6 we implemented a feature where the user could provide a custom
username. The reason for adding this feature was because of
@zanona's very good reason for adding this feature:

> Perhaps we could pass an argument to the generation function in
> order to access an inbox for an email we created in the past?
> Let's say, a login for a forum

But #6 had a minor issue which went unoticed. It only allowed you
provide a username and not an email address. The issue with this is that
it does not satisfy @zanona's need, which was to access an inbox for an
old email address.
Why was that not possible? Because when you gave the username, the TLD
would still be random.

This scenario explains the issue:
Lets say bob@1secmail.com was randomly generate, and I used this
email to sign up for a forum and receive their confirmation email.
I then randomly generated a new email which resulted in this email
address: alice@1secmail.org
A few days later, I want to access the inbox of bob@1secmail.com, so I
run 'tmpmail --generate bob' which might give me this email:
bob@1secmail.net
Notice how it ends with .net and not .com like the email I wanted to
access.

So this commit intends to fix this minor bug/issue.