reinstate check for w3m
This had been removed here to allow running the tool when w3m
is not installed; a better way would be to move the check until
after options have been parsed and possibly an alternate browser
has been requested.  But that would be a different change set /
PR, so putting this back here for now.