Made all variable names lowercase
The reason for doing this is to prevent over writing environment
variables such as $PATH. So in this script would instead use $path as
this variable already does not exist. We currently do not have any
variables that conflict with environment variable but keeping everything
lowercase should hopefully prevent any issues in the future.