Reduce jq(1) invocations to improve speed
When listing the mails of an inbox, the script used 1 jq(1) plus 4 more for each mail. When displaying the contents of a mail, jq(1) was called 4 or 5 times plus 1 for every attachment.

This reduces the amount of calls to 1, except when printing a mail with attachments in raw_text mode, in which case there will be 2 calls. That should greatly speed things up, when printing a large inbox or a mail with many attachments.